refactor(physics+structure): reshuffle modules into engine/structure/physics/analysis
- Move `structure_manipulator/tools.py` -> `structure/transform.py`
  (rename `rattle_structure` -> `rattle`).
- Move `structure_manipulator/interstitials.py` -> `structure/defects.py`
  (rename `substitutional_swap_one_site` -> `substitutional_swap`).
- Move `bulk_defect/vacancy.py:create_vacancy_structure` ->
  `structure/defects.py:create_vacancy`.
- New `structure/build.py`: `get_bulk` (was `get_bulk_structure`) and
  `create_surface_slab` (was `create_surface`).
- Move `bulk.py` -> `physics/bulk.py` (workflow macros only).
- Move `surface/surface_study.py` -> `physics/surface.py`; delete
  `surface/builder.py` (its helpers are in `structure/build.py` now).
- Consolidate `bulk_defect/{point,vacancy}.py` -> `physics/point_defect.py`
  and add the new `get_substitutional_formation_energy` macro.
- Consolidate `gb/{gb_study,cleavage,segregation,optimiser}.py` ->
  `physics/grain_boundary.py`. Internal helpers move to
  `physics/_grain_boundary_helpers/`; `gb/gb_code/` becomes
  `physics/_grain_boundary_code/`.
- Move `gb/analysis.py` -> `analysis/gb_plane.py` (rename
  `find_GB_plane`/`plot_GB_plane` to snake_case).
- Move `featurisers.py` -> `analysis/featurisers.py` (rename the three
  public featuriser functions camelCase -> snake_case).
- Move `utils.py:get_per_atom_quantity` -> `analysis/quantities.py`.
- Update every internal import accordingly. No behaviour change.
