+++ pyiron_workflow_atomistics/analysis/__init__.py
@@ -1,0 +1,21 @@
+"""Featurisation, post-processing, and derived quantities."""
+from .featurisers import (
+    distance_matrix_site_featuriser,
+    pca_whiten,
+    soap_site_featuriser,
+    summarize_cosine_groups,
+    voronoi_site_featuriser,
+)
+from .gb_plane import find_gb_plane, plot_gb_plane
+from .quantities import get_per_atom_quantity
+
+__all__ = [
+    "voronoi_site_featuriser",
+    "distance_matrix_site_featuriser",
+    "soap_site_featuriser",
+    "summarize_cosine_groups",
+    "pca_whiten",
+    "find_gb_plane",
+    "plot_gb_plane",
+    "get_per_atom_quantity",
+]
