--- pyiron_workflow_atomistics/bulk_defect/point.py
@@ -1,1 +1,0 @@
-"""Deprecated stub — symbols moved to physics/point_defect.py in commit 3."""
